convert actionButton to tsx
